+++ modules/logging/core/src/sap/commerce/toolset/logging/state/CxCustomLoggerTemplateState.kt
@@ -21,21 +21,27 @@ package sap.commerce.toolset.logging.state
 import com.intellij.openapi.observable.properties.AtomicProperty
 import com.intellij.openapi.observable.properties.ObservableMutableProperty
 import com.intellij.util.xmlb.annotations.OptionTag
+import java.util.*
 
 data class CxCustomLoggerTemplateState(
+    @OptionTag val uuid: String = UUID.randomUUID().toString(),
     @OptionTag val name: String = "",
     @OptionTag val loggers: List<CxCustomLoggerConfig> = emptyList()
 ) {
+
     fun mutable() = Mutable(
+        uuid = uuid,
         name = AtomicProperty(name),
         loggers = AtomicProperty(loggers.map { it.mutable() }),
     )
 
     data class Mutable(
+        val uuid: String = UUID.randomUUID().toString(),
         val name: ObservableMutableProperty<String>,
         val loggers: ObservableMutableProperty<List<CxCustomLoggerConfig.Mutable>>,
     ) {
         fun immutable() = CxCustomLoggerTemplateState(
+            uuid = uuid,
             name = name.get(),
             loggers = loggers.get().map { it.immutable() },
         )