+++ modules/shared/core/src/sap/commerce/toolset/HybrisIcons.kt
@@ -531,6 +531,7 @@ object HybrisIcons {
         object Action {
             val FETCH = AllIcons.Vcs.Fetch
             val REFRESH = Actions.REFRESH
+            val DELETE = AllIcons.General.Delete
         }
 
         object Template {
@@ -540,8 +541,9 @@ object HybrisIcons {
             val EXECUTE = AllIcons.Actions.Execute
             val ENABLE = getIcon("/icons/logging/template/enableLoggersTemplate.svg")
             val CUSTOM_TEMPLATE = getIcon("/icons/logging/template/customLoggersTemplate.svg")
-            val ADD_CUSTOM_TEMPLATE = getIcon("/icons/logging/template/addCustomLoggersTemplate.svg")
-            val DELETE_CUSTOM_TEMPLATE = getIcon("/icons/logging/template/deleteCustomLoggersTemplate.svg")
+            val ADD_CUSTOM_TEMPLATE = AllIcons.General.Add
+            val DELETE_CUSTOM_TEMPLATE = AllIcons.General.Delete
+            val EDIT_CUSTOM_TEMPLATE = AllIcons.General.Inline_edit
             val DISABLE = AllIcons.Nodes.Template
             val DEFAULT = AllIcons.Nodes.Template
         }